services/PdfService/requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
weasyprint==62.3
jinja2==3.1.4
boto3==1.35.31
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
opentelemetry-api==1.27.0
opentelemetry-sdk==1.27.0
opentelemetry-exporter-otlp==1.27.0
opentelemetry-instrumentation-fastapi==0.48b0
grpcio==1.66.2
protobuf==5.28.2
